ch(auth): test github oauth using netlify
