+++ components/profile/Education/EducationList.tsx
@@ -1,13 +1,13 @@
 "use client";
 import React from "react";
 
-import PaginationContainer from "../../layout/PaginationContainer";
+import PaginationContainer from "@/components/layout/PaginationContainer";
+import NothingtoShow from "@/components/layout/NothingtoShow";
 import EducationCard from "./EducationCard";
 import EducationCardSkeleton from "./EducationCardSkeleton";
-import NothingtoShow from "../../layout/NothingtoShow";
 
 import { useSearchParams } from "next/navigation";
-import { getEducationAction } from "@/utils/actions/education";
+import { getEducationAction } from "@/utils/actions/";
 import { useQuery } from "@tanstack/react-query";
 
 const EducationList = () => {
