+++ components/profile/WorkExperience/index.tsx
@@ -1,11 +1,12 @@
 import React from "react";
-import { useSearchParams } from "next/navigation";
-import WorkExperienceList from "./WorkExperienceList";
 import Link from "next/link";
+import { useSearchParams } from "next/navigation";
+
+import WorkExperienceList from "./WorkExperienceList";
+import { Plus } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import ProfileNavigation from "../ProfileNavigation";
-import { Plus } from "lucide-react";
 
 const WorkExperience = () => {
   const searchParams = useSearchParams();
