+++ utils/actions/developer.ts
@@ -1,94 +1,12 @@
 "use server";
-import { redirect } from "next/navigation";
 import {
   developerFormSchema,
   DeveloperFormType,
   DeveloperType,
   GithubDeveloperRepositoriesType,
-  GithubDeveloperType,
-  loginFormSchema,
-  LoginFormType,
-  signupFormSchema,
-  SignupFormType,
-} from "../types/developer";
-import auth from "../auth";
-
-const getDeveloper = async () => {
-  const developer: DeveloperType | null = await auth.getDeveloper();
-  return developer;
-};
-
-const authenticateAndRedirect = async () => {
-  const developer = await getDeveloper();
-
-  if (!developer || (developer && !developer?.name)) {
-    redirect("/");
-  }
-  return developer;
-};
-
-// signup
-const developerSignupAction = async (
-  values: SignupFormType,
-): Promise<{
-  data?: unknown;
-  error?: string;
-}> => {
-  try {
-    // values validation
-    signupFormSchema.parse(values);
-    // github email validation
-    const suspectedGithubUser: GithubDeveloperType | null =
-      await auth.getGithubDeveloperProfile(values.email);
-
-    if (suspectedGithubUser?.total_count === 0) {
-      // user is not a valid github user
-      return {
-        error: "Signup failed. Please use a valid PUBLIC GITHUB EMAIL.",
-      };
-    }
-    await auth.developerSignup(values);
-    const loginResponse = await auth.developerLogin({
-      email: values.email,
-      password: values.password,
-    });
-    return { data: loginResponse };
-  } catch (error) {
-    console.error(error);
-    return { error: "Something went wrong." };
-  }
-};
-
-// login
-const developerLoginAction = async (
-  values: LoginFormType,
-): Promise<{
-  data?: any;
-  error?: string;
-}> => {
-  try {
-    loginFormSchema.parse(values);
-    const response = await auth.developerLogin(values);
-    return { data: response };
-  } catch (error) {
-    console.error(error);
-    return { error: "Wrong credentails. Login failed." };
-  }
-};
-
-// developer logout
-const developerLogoutAction = async (): Promise<{
-  data?: null;
-  error?: string;
-}> => {
-  try {
-    await auth.deleteSession();
-    return { data: null };
-  } catch (error) {
-    console.error(error);
-    return { error: "Logout failed." };
-  }
-};
+} from "../types/";
+import auth from "../appwrite/auth";
+import { authenticateAndRedirect, getAuthenticatedDeveloper } from "./auth";
 
 // get all developers using a cloud function
 const getDevelopersAction = async (
@@ -105,7 +23,7 @@ const getDevelopersAction = async (
     let usersData = userDocuments.users;
     if (noLoggedInUser) {
       // remove logged in developer from list
-      const developer = await getDeveloper();
+      const developer = await getAuthenticatedDeveloper();
       if (developer?.$id) {
         usersData = userDocuments.users.filter(
           (user: DeveloperType) => user.$id !== developer.$id,
@@ -133,6 +51,7 @@ const updateDeveloper = async (
         developer.$id as string,
         values.password,
       ));
+
     return { data: values.name };
   } catch (error) {
     console.error(error);
@@ -141,7 +60,7 @@ const updateDeveloper = async (
 };
 
 // get developer repositories
-const getGithubDeveloperRepositories = async ({
+const getDeveloperGithubRepositories = async ({
   email,
 }: {
   email: string;
@@ -167,7 +86,7 @@ const getGithubDeveloperRepositories = async ({
 
       if (githubDeveloperProfileName) {
         const githubDeveloperRepositoriesResponse =
-          await auth.getGithubDeveloperRepositories(githubDeveloperProfileName);
+          await auth.getDeveloperGithubRepositories(githubDeveloperProfileName);
 
         if (githubDeveloperRepositoriesResponse) {
           githubDeveloperRepositories = githubDeveloperRepositoriesResponse.map(
@@ -193,12 +112,7 @@ const getGithubDeveloperRepositories = async ({
 
 export {
   // developer
-  authenticateAndRedirect,
-  getDeveloper,
   getDevelopersAction,
-  developerSignupAction,
-  developerLoginAction,
-  developerLogoutAction,
   updateDeveloper,
-  getGithubDeveloperRepositories,
+  getDeveloperGithubRepositories,
 };
