+++ utils/magicValues.ts
@@ -1,7 +1,1 @@
 export const DEFAULT_PAGE_LIMIT = 10;
-export const PASSWORD_REGEX = new RegExp(
-  "^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])(?=.{8,})",
-);
-// new RegExp(
-//   /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/
-// );
