+++ utils/types/comments.ts
@@ -1,6 +1,6 @@
 import * as z from "zod";
 
-// POST COMMENTS
+// POST COMMENTS TYPES
 export type PostCommentType = {
   $id?: string;
   $createdAt: string;
@@ -13,18 +13,11 @@ export type PostCommentType = {
   mine: boolean;
 };
 
+export type PostCommentFormType = z.infer<typeof postCommentFormSchema>;
+
+// POST COMMENTS zod infer type
 export const postCommentFormSchema = z.object({
   comment: z.string().min(2, {
     message: "Field must be at least 2 characters.",
   }),
 });
-
-export type PostCommentFormType = z.infer<typeof postCommentFormSchema>;
-
-// POST LIKES/DISLIKES
-export type PostLikeType = {
-  $id?: string;
-  $createdAt: string;
-  $updatedAt: string;
-  developerId: string;
-};
