+++ utils/types/likes.ts
@@ -1,0 +1,7 @@
+// POST LIKES/DISLIKES
+export type PostLikeType = {
+  $id?: string;
+  $createdAt: string;
+  $updatedAt: string;
+  developerId: string;
+};
