+++ utils/types/posts.ts
@@ -1,7 +1,7 @@
 import * as z from "zod";
-import { PostCommentType } from "./postLikesComments";
+import { PostCommentType } from "./comments";
 
-// POSTS
+// POSTS TYPES
 export type PostType = {
   $id: string;
   $createdAt: string;
@@ -18,10 +18,11 @@ export type PostType = {
   comments?: PostCommentType[];
 };
 
+export type PostFormType = z.infer<typeof postFormSchema>;
+
+// POSTS SCHEMA
 export const postFormSchema = z.object({
   message: z.string().min(2, {
     message: "Field must be at least 2 characters.",
   }),
 });
-
-export type PostFormType = z.infer<typeof postFormSchema>;
